backend/requirements.txt
bcrypt==4.1.3
black>=24.1.1
boto3>=1.34.129
cryptography>=42.0.8
email-validator>=2.2.0
emergentintegrations==0.1.0
fastapi==0.110.1
flake8>=7.0.0
httpx>=0.27.0
isort>=5.13.2
jq>=1.6.0
motor==3.3.1
mypy>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
passlib>=1.7.4
pillow==12.1.0
pydantic>=2.6.4
pyjwt>=2.10.1
pymongo==4.5.0
pytest>=8.0.0
python-dotenv>=1.0.1
python-jose>=3.3.0
python-multipart>=0.0.9
requests-oauthlib>=2.0.0
requests>=2.31.0
typer>=0.9.0
tzdata>=2024.2
uvicorn==0.25.0
nest-asyncio==1.6.0
